+++ modules/kubernetes/cluster-autoscaler/main.tf
@@ -16,20 +16,30 @@ terraform {
       source  = "hashicorp/helm"
       version = "2.5.1"
     }
+    kubectl = {
+      source  = "gavinbunney/kubectl"
+      version = "1.14.0"
+    }
   }
 }
 
 locals {
-  namespace = "cluster-autoscaler"
+  tags = {
+    "1.22" = "1.22.3"
+    "1.23" = "1.23.1"
+    "1.24" = "1.24.0"
+  }
 }
+
+data "kubectl_server_version" "current" {}
 
 resource "kubernetes_namespace" "this" {
   metadata {
+    name = "cluster-autoscaler"
     labels = {
-      name                = local.namespace
+      name                = "cluster-autoscaler"
       "xkf.xenit.io/kind" = "platform"
     }
-    name = local.namespace
   }
 }
 
@@ -38,11 +48,12 @@ resource "helm_release" "cluster_autoscaler" {
   chart       = "cluster-autoscaler"
   name        = "cluster-autoscaler"
   namespace   = kubernetes_namespace.this.metadata[0].name
-  version     = "9.10.7"
+  version     = "9.19.2"
   max_history = 3
   values = [templatefile("${path.module}/templates/values.yaml.tpl", {
     provider     = var.cloud_provider,
     cluster_name = var.cluster_name
     aws_config   = var.aws_config
+    tag          = local.tags["${data.kubectl_server_version.current.major}.${data.kubectl_server_version.current.major}"]
   })]
 }